Separating classes (draft)
This is just a draft. I didnt finish what I was doing yet
